frontend: optimisation of list and map with cache.
