+++ main/local_modules/frontend_base/controllers/base.py
@@ -3,16 +3,33 @@ import logging
 
 import simplejson
 from datadog import statsd
-from openerp.addons.frontend_base.models.caches import caches
 from openerp.addons.web import http
 from openerp.addons.website.controllers.main import Website
+from openerp.addons.website.models.website import hashlib
 
 from openerp.http import request
 
 _logger = logging.getLogger(__name__)
 
-cache = caches.get('cache_3h')
 
+def small_image_url(record, field):
+    """Returns a local url that points to the image field of a given browse record."""
+    if not record.small_image_url:
+        _logger.debug('No small image url for %s', record.id)
+        model = record._name
+        sudo_record = record.sudo()
+        id_ = '%s_%s' % (
+            record.id,
+            hashlib.sha1(
+                sudo_record.write_date or sudo_record.create_date or ''
+            ).hexdigest()[0:7]
+        )
+        size = '' if size is None else '/%s' % size
+        record.small_image_url = '/website/image/%s/%s/%s%s' % (model, id_, field, size)
+        # else:
+        # _logger.debug('Great found small image url for %s!', record.id)
+
+    return record.small_image_url
 
 class Base(Website):
     """Representation of the homepage of the website."""
