+++ main/local_modules/frontend_base/models/__init__.py
@@ -24,6 +24,5 @@ from . import (
     res_partner,
     res_users,
     website,
-    caches,
     res_industry
 )
